+++ examples/random_dungeon.rs
@@ -50,17 +50,17 @@ impl GameState {
 }
 
 fn setup(
-    mut commands: Commands,
+    commands: &mut Commands,
     mut tile_sprite_handles: ResMut<TileSpriteHandles>,
     asset_server: Res<AssetServer>,
 ) {
     tile_sprite_handles.handles = asset_server.load_folder("textures").unwrap();
 
-    commands.spawn(Camera2dComponents::default());
+    commands.spawn(Camera2dBundle::default());
 }
 
 fn load(
-    mut commands: Commands,
+    commands: &mut Commands,
     mut sprite_handles: ResMut<TileSpriteHandles>,
     mut texture_atlases: ResMut<Assets<TextureAtlas>>,
     mut textures: ResMut<Assets<Texture>>,
@@ -93,7 +93,7 @@ fn load(
             .finish()
             .unwrap();
 
-        let tilemap_components = TilemapComponents {
+        let tilemap_components = TilemapBundle {
             tilemap,
             transform: Default::default(),
             global_transform: Default::default(),
@@ -108,7 +108,7 @@ fn load(
 }
 
 fn build_random_dungeon(
-    mut commands: Commands,
+    commands: &mut Commands,
     mut game_state: ResMut<GameState>,
     texture_atlases: Res<Assets<TextureAtlas>>,
     asset_server: Res<AssetServer>,
@@ -241,6 +241,7 @@ fn move_sprite(
 fn character_movement(
     mut game_state: ResMut<GameState>,
     keyboard_input: Res<Input<KeyCode>>,
+    time: Res<Time>,
     mut map_query: Query<(&mut Tilemap, &mut Timer)>,
     mut player_query: Query<(&mut Position, &Render, &Player)>,
 ) {
@@ -248,8 +249,9 @@ fn character_movement(
         return;
     }
 
-    for (mut map, timer) in map_query.iter_mut() {
-        if !timer.finished {
+    for (mut map, mut timer) in map_query.iter_mut() {
+        timer.tick(time.delta_seconds());
+        if !timer.finished() {
             continue;
         }
 
@@ -299,8 +301,8 @@ fn main() {
     App::build()
         .add_resource(WindowDescriptor {
             title: "Random Tile Dungeon".to_string(),
-            width: 1024,
-            height: 1024,
+            width: 1024.,
+            height: 1024.,
             vsync: false,
             resizable: false,
             mode: WindowMode::Windowed,